Refector Collection db util to seperate class
