+++ src/main/java/de/samply/bbmri/negotiator/control/UserBean.java
@@ -65,6 +65,7 @@ import de.samply.common.config.OAuth2Client;
 import de.samply.string.util.StringUtil;
 //import org.slf4j.Logger;
 //import org.slf4j.LoggerFactory;
+import eu.bbmri.eric.csit.service.negotiator.database.DbUtilCollection;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -503,7 +504,7 @@ public class UserBean implements Serializable {
 			/**
 			 * Check if the user is a biobanker
 			 */
-            collections = DbUtil.getCollections(config, person.getId());
+            collections = DbUtilCollection.getCollections(config, person.getId());
             biobankOwner = collections.size() > 0;
 
             // Check if user is NationalNodeRepresentative
